11/08
Today I mostly build more of the interface. Also, the scene management in the previous commit was just based on function nesting. So now I decided to make one that doesn't do that and instead use branching of the main function. I plan on writing a rust code to automate this process in the future so I don't need to have a scene table like the one I currently have.